Фикс с удалением элементов из блоков
