+++ letter-builder/atoms/LineBlocksContent/Button/Button.tsx
@@ -1,38 +1,13 @@
-import { useAppDispatch } from "@store/store";
-import { addElement, initPanel, updateText } from "@/store/LetterBuilderStore/styleModule";
-import { useAppSelector } from "@/store/store";
-import { useEffect } from "react";
+import { useStyleElement } from "../../../hooks/useStyleElement";
 
 const ButtonComponent = ({ id }: { id: string }) => {
-  const dispatch = useAppDispatch();
-  const parameters = useAppSelector((state) => state.styleModule.elements?.[id]);
-
-  useEffect(() => {
-    if (!parameters) {
-      dispatch(
-        addElement({
-          id,
-          styles: {
-            backgroundColor: "#2563eb",
-            color: "#ffffff",
-            borderRadius: "6px",
-            fontSize: "14px",
-            lineHeight: "20px",
-          },
-          type: "button",
-        }),
-      );
-    }
-  }, [parameters, dispatch, id]);
-
-  const handleOpen = (event: React.MouseEvent) => {
-    event.preventDefault();
-    dispatch(initPanel(id));
-  };
-
-  const handleTextChange = (text: string) => {
-    dispatch(updateText({ id, text }));
-  };
+  const { handleOpen, handleTextChange, parameters } = useStyleElement(id, {
+    backgroundColor: "#2563eb",
+    color: "#ffffff",
+    borderRadius: "6px",
+    fontSize: "14px",
+    lineHeight: "20px",
+  });
 
   return (
     <a
@@ -42,6 +17,7 @@ const ButtonComponent = ({ id }: { id: string }) => {
         transition: "all 0.2s ease-in-out",
         borderStyle: "solid",
         borderWidth: "1px",
+        outline: "none",
         ...parameters?.styles,
       }}
       contentEditable
