+++ letter-builder/atoms/StyleAccordion/index.ts
@@ -1,0 +1,1 @@
+export { StyleAccordion } from "./style-accordion";
